requirements.txt
pandas>=2.0
python-dotenv>=1.0
requests>=2.31
websockets>=11.0
alpaca-py>=0.13